requirements.txt
streamlit
pdfplumber
camelot-py[cv]
pandas
requests
openpyxl
xlrd
numpy